Revert "Merge pull request #21 from Finefinee/invite3"
This reverts commit 17bf9a1e23a89ec79f0fa1980fa96e3b027c8a7e, reversing
changes made to d4c311f501aa202bfbd8e6f98d0abc330243a185.

+++ src/main/java/Project/Teaming/Invite/Dto/AcceptInviteRequestDto.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class AcceptInviteRequestDto {
 
-    private Integer inviteId;
+    private Long inviteId;
 }
 
 // test committ
