+++ src/main/java/Project/Teaming/Invite/Repository/InviteRepository.java
@@ -3,6 +3,6 @@ package Project.Teaming.Invite.Repository;
 import Project.Teaming.Invite.Entity.Invite;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InviteRepository extends JpaRepository<Invite, Integer> {
+public interface InviteRepository extends JpaRepository<Invite, Long> {
 
 }
